implement Redux into the project (#22)
* implement Redux into the project

* implement Redux into the project

fix after review

+++ src/views/mainBoard/board/column/Column.tsx
@@ -12,23 +12,24 @@ import { ColorPicker } from "antd";
 import { Color } from "antd/es/color-picker";
 import DeleteModal from "../../../modals/deleteBoardOrTaskModal/DeleteModal";
 import { useState } from "react";
+import { useDispatch } from "react-redux";
+import { deleteTask, updateColumn } from "../../../../state/selectedBoard/selectedBoardSlice";
+import { AppDispatch } from "../../../../state/store";
 
 interface ColumnProps {
     column: ColumnType;
-    statusOptions: string[];
-    updateColumn: (column: ColumnType) => void;
-    updateColumnsAfterTaskStatusChange: (task: TaskType, columnId: string) => void;
 }
 
-const Column = ({ column, statusOptions, updateColumn, updateColumnsAfterTaskStatusChange }: ColumnProps) => {
-    const [isDeleteModalopen, setIsDeleteModalOpen] = useState(false);
+const Column = ({ column }: ColumnProps) => {
+    const dispatch = useDispatch<AppDispatch>();
+    const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
     const [taskToDelete, setTaskToDelete] = useState<TaskType | null>(null);
     const numberOfTasks = column.tasks?.length;
     const deleteModalText = taskToDelete ? getDeleteModalText(taskToDelete.title) : "";
 
     const handleColumnColorChange = (color: Color) => {
         const updatedColumn = { ...column, color: color.toHexString() };
-        updateColumn(updatedColumn);
+        dispatch(updateColumn(updatedColumn));
     };
 
     const handleDeleteModalOpen = (task: TaskType) => {
@@ -41,25 +42,13 @@ const Column = ({ column, statusOptions, updateColumn, updateColumnsAfterTaskSta
     };
 
     const handleDeleteTask = () => {
-        const updatedTasks = column.tasks.filter((task) => task.id !== taskToDelete?.id);
-        const updatedColumn = { ...column, tasks: updatedTasks };
-        updateColumn(updatedColumn);
+        if (!taskToDelete) {
+            return;
+        }
+
+        dispatch(deleteTask({ task: taskToDelete, columnId: column.id }));
         setTaskToDelete(null);
         handleDeleteModalCancel();
-    };
-
-    const handleEditTask = (updatedTask: TaskType, previousStatus?: string) => {
-        if (previousStatus && previousStatus !== updatedTask.status) {
-            updateColumnsAfterTaskStatusChange(updatedTask, column.id);
-        } else {
-            const updatedTasks = column.tasks.map((task) => (task.id === updatedTask.id ? updatedTask : task));
-            const updatedColumn = { ...column, tasks: updatedTasks };
-            updateColumn(updatedColumn);
-        }
-    };
-
-    const handleChangeTaskStatus = (updatedTask: TaskType) => {
-        updateColumnsAfterTaskStatusChange(updatedTask, column.id);
     };
 
     //TODO: handle dark mode
@@ -68,7 +57,7 @@ const Column = ({ column, statusOptions, updateColumn, updateColumnsAfterTaskSta
         <>
             <StyledMainContainer>
                 <StyledContainer>
-                    <ColorPicker onChangeComplete={handleColumnColorChange} defaultValue={column.color}>
+                    <ColorPicker onChangeComplete={handleColumnColorChange} value={column.color}>
                         <StyledColumnIcon style={{ backgroundColor: column.color }} />
                     </ColorPicker>
                     <StyledTitle>
@@ -80,17 +69,15 @@ const Column = ({ column, statusOptions, updateColumn, updateColumnsAfterTaskSta
                         <Task
                             key={task.id}
                             task={task}
-                            statusOptions={statusOptions}
-                            editTask={handleEditTask}
                             deleteModalOpen={handleDeleteModalOpen}
-                            changeTaskStatus={handleChangeTaskStatus}
+                            updatedColumnId={column.id}
                         />
                     ))}
                 </StyledTasksContainer>
             </StyledMainContainer>
 
             <DeleteModal
-                open={isDeleteModalopen}
+                open={isDeleteModalOpen}
                 onCancel={handleDeleteModalCancel}
                 onDelete={handleDeleteTask}
                 title={deleteModalTitle}
